bandmean: use 64-bits intermediate for int format

+++ libvips/conversion/bandmean.c
@@ -148,7 +148,7 @@ vips_bandmean_buffer(VipsBandarySequence *seq,
 		UILOOP(unsigned short, unsigned int);
 		break;
 	case VIPS_FORMAT_INT:
-		SILOOP(signed int, int);
+		SILOOP(signed int, int64_t);
 		break;
 	case VIPS_FORMAT_UINT:
 		UILOOP(unsigned int, unsigned int);
